Move auth socket waiter out of TerminalController
